add a function to merge mp4 files based on media3 for Android

+++ packages/nativescript-transcoder/index.android.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
-import { File, Utils } from '@nativescript/core';
+import { Device, File, Utils } from '@nativescript/core';
 import { NativescriptTranscoderCommon, VideoConfig, VideoResolution } from './common';
 import { clearInterval, setInterval } from '@nativescript/core/timer';
 
@@ -14,6 +14,11 @@ export class NativescriptTranscoder extends NativescriptTranscoderCommon {
    */
   transcode(inputPath: string, outputPath: string, videoConfig: VideoConfig): Promise<File> {
     return new Promise((resolve, reject) => {
+      if (+Device.sdkVersion < 21) {
+        console.error('This is only supported on API 21+');
+        return reject('This is only supported on API 21+');
+      }
+
       if (File.exists(outputPath)) {
         const file = File.fromPath(outputPath);
         file.removeSync();
@@ -118,7 +123,7 @@ export class NativescriptTranscoder extends NativescriptTranscoderCommon {
   }
 
   /**
-   * Transcodes audio from inputPath to outputPath as an MP4 container with AAC audio
+   *  **ANDROID-ONLY** Transcodes the audio input file to an MP4 file with AAC encoding
    * @param inputPath string (can be a url or absolute path)
    * @param outputPath string
    * @returns Promise<File>
@@ -126,6 +131,11 @@ export class NativescriptTranscoder extends NativescriptTranscoderCommon {
    */
   convertAudioToMp4(inputPath: string, outputPath: string): Promise<File> {
     return new Promise((resolve, reject) => {
+      if (+Device.sdkVersion < 21) {
+        console.error('This is only supported on API 21+');
+        return reject('This is only supported on API 21+');
+      }
+
       if (File.exists(outputPath)) {
         const file = File.fromPath(outputPath);
         file.removeSync();
@@ -153,6 +163,7 @@ export class NativescriptTranscoder extends NativescriptTranscoderCommon {
 
       // eslint-disable-next-line @typescript-eslint/no-this-alias
       const that = this;
+
       const listener: androidx.media3.transformer.Transformer.Listener = new androidx.media3.transformer.Transformer.Listener({
         onTransformationCompleted: (inputMediaItem: androidx.media3.common.MediaItem) => {
           that.log('onTransformationCompleted');
@@ -201,6 +212,111 @@ export class NativescriptTranscoder extends NativescriptTranscoderCommon {
         this.log('Progress: ' + progressHolder.progress / 100);
         emit(NativescriptTranscoderCommon.TRANSCODING_PROGRESS, { progress: progressHolder.progress / 100 });
       }, 200);
+    });
+  }
+
+  /**
+   * Merges the mp4 files specified by inputFiles (array of file paths) into an mp4 file at the outputPath.
+   *
+   * On iOS,  input files must all have audio and video tracks with the same codecs used.
+   * @method mergeMp4Files
+   * @param inputFiles
+   * @param outputPath
+   * @returns Promise<File>
+   **/
+  public mergeMp4Files(inputFiles: string[], outputPath: string): Promise<File> {
+    return new Promise((resolve, reject) => {
+      if (+Device.sdkVersion < 21) {
+        console.error('This is only supported on API 21+');
+        return reject('This is only supported on API 21+');
+      }
+      if (!inputFiles || inputFiles.length <= 0) return reject('inputFiles is empty!');
+      if (!outputPath) return reject('outputPath should be a valid path string');
+      if (File.exists(outputPath)) {
+        // remove file if it exists
+        File.fromPath(outputPath).removeSync(err => {
+          console.error('Unable to remove file at output path!', err);
+          return reject('Unable to remove file at output path!' + err.message);
+        });
+      }
+      if (inputFiles.length == 1) {
+        const fileData = File.fromPath(inputFiles[0]).readSync();
+        File.fromPath(outputPath).writeSync(fileData);
+        return resolve(File.fromPath(outputPath));
+      }
+      const emit = (event: string, data: any) => {
+        this.notify({ eventName: event, object: this, data });
+      };
+
+      //create the list of EditedMediaItem from file paths to add to the composition
+      const segments: java.util.List<androidx.media3.transformer.EditedMediaItem> = new java.util.ArrayList();
+      try {
+        for (let i = 0; i < inputFiles.length; i++) {
+          //add a EditedMediaItem for each file to be merged
+          const inputMediaItem: androidx.media3.common.MediaItem = androidx.media3.common.MediaItem.fromUri(inputFiles[i]);
+          const editedMediaItem: androidx.media3.transformer.EditedMediaItem = new androidx.media3.transformer.EditedMediaItem.Builder(inputMediaItem).build();
+          segments.add(editedMediaItem);
+        }
+
+        const sequence: androidx.media3.transformer.EditedMediaItemSequence = new androidx.media3.transformer.EditedMediaItemSequence(segments);
+        const sequenceList = new java.util.ArrayList();
+        sequenceList.add(sequence);
+        const composition: androidx.media3.transformer.Composition = new androidx.media3.transformer.Composition.Builder(sequenceList).build();
+
+        // eslint-disable-next-line @typescript-eslint/no-this-alias
+        const that = this;
+        const listener: androidx.media3.transformer.Transformer.Listener = new androidx.media3.transformer.Transformer.Listener({
+          onTransformationCompleted: (inputMediaItem: androidx.media3.common.MediaItem) => {
+            that.log('onTransformationCompleted');
+            emit(NativescriptTranscoderCommon.TRANSCODING_COMPLETE, {});
+            resolve(File.fromPath(outputPath));
+            clearInterval(progressUpdater);
+          },
+          onCompleted: (composition: androidx.media3.transformer.Composition, exportResult: androidx.media3.transformer.ExportResult) => {
+            that.log('onCompleted');
+            emit(NativescriptTranscoderCommon.TRANSCODING_COMPLETE, { output: outputPath });
+            resolve(File.fromPath(outputPath));
+            clearInterval(progressUpdater);
+          },
+          //@ts-ignore
+          onTransformationError: (inputMediaItem: androidx.media3.common.MediaItem, exception: androidx.media3.transformer.TransformationException) => {
+            that.log('onTransformationError!', exception);
+            emit(NativescriptTranscoderCommon.TRANSCODING_ERROR, { error: exception });
+            reject(exception);
+            clearInterval(progressUpdater);
+          },
+          onError: (composition: androidx.media3.transformer.Composition, exportResult: androidx.media3.transformer.ExportResult, exportException: androidx.media3.transformer.ExportException) => {
+            that.log('onError', exportException);
+            emit(NativescriptTranscoderCommon.TRANSCODING_ERROR, { error: exportException });
+            reject(exportException);
+            clearInterval(progressUpdater);
+          },
+          //@ts-ignore
+          onFallbackApplied: (
+            inputMediaItem: androidx.media3.common.MediaItem,
+            originalTransformationRequest: androidx.media3.transformer.TransformationRequest,
+            fallbackTransformationRequest: androidx.media3.transformer.TransformationRequest
+          ) => {
+            this.log('onFallbackApplied');
+          },
+        });
+        const transformer: androidx.media3.transformer.Transformer = new androidx.media3.transformer.Transformer.Builder(this.getAndroidContext())
+          .setVideoMimeType(androidx.media3.common.MimeTypes.VIDEO_H264)
+          .setAudioMimeType(androidx.media3.common.MimeTypes.AUDIO_AAC)
+          .addListener(listener)
+          .build();
+        transformer.start(composition, outputPath);
+        emit(NativescriptTranscoderCommon.TRANSCODING_STARTED, null);
+        const progressHolder: androidx.media3.transformer.ProgressHolder = new androidx.media3.transformer.ProgressHolder();
+        const progressUpdater = setInterval(() => {
+          transformer.getProgress(progressHolder);
+          this.log('Progress: ' + progressHolder.progress / 100);
+          emit(NativescriptTranscoderCommon.TRANSCODING_PROGRESS, { progress: progressHolder.progress / 100 });
+        }, 200);
+      } catch (err) {
+        console.error(err, err.message);
+        return reject('Error during merge: ' + err.message);
+      }
     });
   }
